+++ integration/single.csv
@@ -1,4 +1,4 @@
 index,build,test_index,y_D,run_index
-0,Local,0,4.0,0
-1,Local,0,4.0,1
-2,Local,0,4.0,2
+0,Click 2022,0,4.0,0
+1,Click 2022,0,4.0,1
+2,Click 2022,0,4.0,2
