+++ npf/models/series.py
