+++ npf/output/graph/graphdata.py
@@ -1,5 +1,5 @@
 from ordered_set import OrderedSet
-from npf.types.dataset import convert_to_xyeb
+from npf.models.dataset import convert_to_xyeb
 from npf.tests.variable import is_numeric, get_numeric
 import numpy as np
