+++ npf/tests/build.py
@@ -5,10 +5,10 @@ from subprocess import PIPE
 from pathlib import Path
 import re
 from npf.globals import get_options
-from npf.types.dataset import Run, Dataset
+from npf.models.dataset import Run, Dataset
 import copy
 
-import npf.types.units
+import npf.models.units
 
 renametable = {
     'npf.script': 'npf.npf',
@@ -196,7 +196,7 @@ class Build:
                 for v_data in re.split(r'(?<!\\),', variables_data):
                     if v_data.strip():
                         k, v = re.split(r'(?<!\\):', v_data)
-                        variables[k] = npf.types.units.get_numeric(v) if npf.types.units.is_numeric(k) else str(v).replace('\\:',':')
+                        variables[k] = npf.models.units.get_numeric(v) if npf.models.units.is_numeric(k) else str(v).replace('\\:',':')
                 results = {}
 
                 results_data = results_data.strip()[1:-1].split('},{')
