+++ npf/tests/variable.py
@@ -1,32 +1,21 @@
 import re
 from collections import OrderedDict
 
-import regex
 import random
 from asteval import Interpreter
 import itertools
 import numpy as np
 import csv
-import sys
+
 
 import random
 
-from npf.cluster.nic import NIC
+import npf.cluster.node
+from npf.models.variables.variable import Variable
 import npf.osutils
-from npf.types.units import is_numeric
-from npf.types.units import is_integer
-from npf.types.units import get_bool
-from npf.types.units import is_bool
-from npf.types.units import get_numeric
-
-def dtype(v):
-    if is_numeric(v):
-        if is_integer(v):
-            return int
-        else:
-            return float
-    else:
-        return str
+from npf.models.units import is_numeric
+from npf.models.units import get_bool
+from npf.models.units import get_numeric
 
 def is_log(l):
     if len(l) < 3:
@@ -88,7 +77,7 @@ def replace_variables(v: dict, content: str, self_role=None, self_node=None, def
     def do_replace_nics(nic_match):
         varRole = nic_match.group('role')
 
-        nodes = npf.nodes_for_role(varRole, self_role, self_node, default_role_map)
+        nodes = npf.cluster.factory.nodes_for_role(varRole, self_role, self_node, default_role_map)
         nodeidx = role_index % len(nodes)
         if nic_match.groupdict()['node']:
             t = str(nic_match.group('node'))
@@ -131,69 +120,7 @@ def replace_variables(v: dict, content: str, self_role=None, self_node=None, def
 
 
 
-class VariableFactory:
-    @staticmethod
-    def build(name, valuedata, vsection=None):
-        result = re.match("(?P<doubleopen>\[?)\[(?P<a>-?[0-9.]+)(?P<log>[+-]|[*]|[,])(?P<b>-?[0-9.]+)(?P<step>[#][0-9.]*)?\](?P<doubleclose>\]?)", valuedata)
-        if result:
-            return RangeVariable(name, result.group('a'), result.group('b'), result.group('log') == "*", step= (get_numeric(result.group('step')[1:]) if result.group('step') else None), force_int=result.group('doubleopen')=='[')
-
-        result = regex.match("\{([^:]*:[^,:]+)(?:(?:,)([^,:]*:[^,:]+))*\}", valuedata)
-        if result:
-            return DictVariable(name, result.captures(1) + result.captures(2))
-
-        result = regex.match("\{([^,]+)(?:(?:,)([^,]*))*}", valuedata)
-        if result:
-            return ListVariable(name, result.captures(1) + result.captures(2))
-        if valuedata.strip() == "{}":
-            return DictVariable(name, {})
-
-        result = regex.match("EXPAND\((.*)\)", valuedata)
-        if result:
-            if vsection is None:
-                raise Exception("RANDOM variable without vsection",vsection)
-            return ExpandVariable(name, result.group(1), vsection)
-
-        result = regex.match("RANDOM[ ]*\([ ]*([^,]+)[ ]*,[ ]*([^,]+)[ ]*\)", valuedata)
-        if result:
-            if vsection is None:
-                raise Exception("RANDOM variable without vsection",vsection)
-            return RandomVariable(name, vsection.replace_all(result.group(1))[0], vsection.replace_all(result.group(2))[0])
-
-        result = regex.match("HEAD[ ]*\([ ]*([^,]+)[ ]*,[ ]*\$([^,]+)[ ]*(,[ ]*(?P<sep>.+)[ ]*)?\)", valuedata)
-        if result:
-            if vsection is None:
-                raise Exception("HEAD variable without vsection",vsection)
-            nums = vsection.replace_all(result.group(1))[0].strip()
-            return HeadVariable(name, nums,
-                                vsection.vlist[result.group(2)].makeValues(), result.group('sep'))
-        result = regex.match("IF[ ]*\([ ]*([^,]+)[ ]*,[ ]*([^,]+)[ ]*,[ ]*([^,]+)[ ]*\)", valuedata)
-        if result:
-            if vsection is None:
-                raise Exception("IF variable without vsection",vsection)
-            return IfVariable(name, vsection.replace_all(result.group(1))[0], result.group(2), result.group(3))
-
-        return SimpleVariable(name, valuedata)
-
-
 # raise Exception("Unkown variable type : " + valuedata)
-
-class Variable:
-    def __init__(self, name):
-        self.assign = '='
-        self.is_default = False
-        self.name = name
-
-    NAME_REGEX = r'[a-zA-Z0-9._-]+'
-    TAGS_REGEX = r'[a-zA-Z0-9._,|!-]+'
-    VALUE_REGEX = r'[a-zA-Z0-9._/,{}^$-:]+'
-    VARIABLE_REGEX = r'(?<!\\)[$](' \
-                     r'[{](?P<varname_in>' + NAME_REGEX + ')[}]|' \
-                     r'(?P<varname_sp>' + NAME_REGEX + ')(?=}|[^a-zA-Z0-9_]|$))'
-    MATH_REGEX = r'(?P<prefix>\\)?[$][(][(](?P<expr>.*?)[)][)]'
-    ALLOWED_NODE_VARS = 'path|user|addr|tags|nfs|arch|port'
-    NICREF_REGEX = r'(?P<role>[a-z0-9]+)[:](:?(?P<nic_idx>[0-9]+)[:](?P<type>' + NIC.TYPES + '+)|(?P<node>'+ALLOWED_NODE_VARS+'|ip|ip6|multi|mode|node))'
-    VARIABLE_NICREF_REGEX = r'(?<!\\)[$][{]' + NICREF_REGEX + '[}]'
 
 class FromFileVariable:
     """Deprecated. Use the experimental design system instead of drawing from a file.
@@ -222,353 +149,4 @@ class FromFileVariable:
 
         # TODO: assert that the number of rows of the matrix is sufficient for all the values of the experimental design variables
 
-class CoVariable(Variable):
-    """A serie of variable that should advance together.
-    
-    A=[1-10]
-    B=[1-10]
-    
-    Normally that creates 100 combinations.
-    """
-    def __init__(self, name = "covariable"):
-        super().__init__(name)
-        self.vlist = OrderedDict()
-
-    def makeValues(self):
-        vs = [OrderedDict() for i in range(self.count())]
-        vals = []
-        for i,(var,val) in enumerate(self.vlist.items()):
-            vals.append(val.makeValues())
-        for i in range(self.count()):
-            for j,var in enumerate(self.vlist.keys()):
-                vs[i][var] = vals[j][i]
-        return vs
-
-    def count(self):
-        if len(self.vlist) == 0:
-            return 1
-        return min([v.count() for k,v in self.vlist.items()])
-
-    def format(self):
-        k = []
-        v = []
-        for var,val in self.vlist.items():
-            K,V = val.format()
-            k.append(K)
-            v.append(V)
-        return k,v
-
-    def is_numeric(self):
-        return False
-
 # For each value N of nums, generate a variable with the first N element of values
-class HeadVariable(Variable):
-    """DEPRECATED, use jinja instead.
-    """
-    def __init__(self, name, nums, values, join = None):
-        super().__init__(name)
-        self.values = values
-        if not is_numeric(nums):
-            raise Exception("%s is not a number!" % nums)
-        self.nums = nums
-        self.join = join if join else "\n"
-
-
-    def makeValues(self):
-        if self.nums == 0:
-            return ['']
-        vs = []
-        i = int(self.nums)
-        if type(i) is str:
-            i = int(i.strip())
-        vs.append((self.join.join(self.values[:i]), i))
-        return vs
-
-    def count(self):
-        return sum(self.nums) if len(self.nums) > 0 else 1
-
-    def format(self):
-        return self.name, str
-
-    def is_numeric(self):
-        return False
-
-class ExpandVariable(Variable):
-    """ Create a list which expands a string with all possible value for the variable
-        it contains like it would be in a script or file section.
-        
-        This can be seen as a variable that will interpret all variables inside it. For instance
-        A=7
-        B=EXPAND(SYSTEM-$A)
-        
-        --> A=7, B=SYSTEM-7
-        
-        However if A has multiple values, it will also create multiple B.
-        
-        This should be avoided. The good practice is to use a jinja template in the test itself.
-        
-        A=7
-        
-        %script@dut jinja
-        echo "SYSTEM-{{A}}"
-        """
-    def __init__(self, name, value, vsection):
-        super().__init__(name)
-        self.values = vsection.replace_all(value)
-
-    def makeValues(self):
-        return self.values
-
-    def count(self):
-        return len(self.values)
-
-    def format(self):
-        return self.name, str
-
-    def is_numeric(self):
-        return False
-
-    def __add__(self, other):
-        v = []
-        for ov in other.makeValues():
-            for mv in self.values:
-                v.append(mv + ov)
-        self.values = v
-        return self
-
-class SimpleVariable(Variable):
-    """Variable with just a value
-    """
-    def __init__(self, name, value):
-        super().__init__(name)
-        self.value = get_numeric(value)
-
-    def makeValues(self):
-        return [self.value]
-
-    def count(self):
-        return 1
-
-    def format(self):
-        return self.name, dtype(self.value)
-
-    def is_numeric(self):
-        return self.format() != str
-
-    def __add__(self, other):
-        self.value += other.makeValues()[0]
-        return self
-
-    def pop(self, item):
-        if self.value == item:
-            return ListVariable(None, [])
-        else:
-            return self
-
-
-class ListVariable(Variable):
-    """A variable with a list of elements
-    """
-    def __init__(self, name, l):
-        super().__init__(name)
-        all_num = True
-        for x in l:
-            if not is_numeric(x):
-                all_num = False
-                break
-        if all_num:
-            self.lvalues = [int(x) if is_integer(x) else float(x) if is_numeric(x) else x for x in l]
-        else:
-            self.lvalues = [str(x) for x in l]
-        self.all_num = all_num
-
-    def __add__(self,other):
-        self.lvalues.extend(other.lvalues)
-        return self
-
-    def makeValues(self):
-        vs = []
-        for v in self.lvalues:
-            if (v is None):
-                continue
-            vs.append(v)
-        return vs
-
-    def count(self):
-        return len(self.lvalues)
-
-    def format(self):
-        t = dtype(self.lvalues[0])
-        if t is int and is_bool(self.lvalues[0]):
-            unique = list(set(self.lvalues))
-            if len(unique) == 2 and all([int(u) in (0, 1) for u in unique]):
-                return self.name, bool
-            elif len(unique) == 1 and (unique[0] == 'true' or unique[0] == 'false'):
-                return self.name, bool
-        return self.name, t
-
-    def is_numeric(self):
-        return self.all_num
-
-    def pop(self, item):
-        if item in self.lvalues:
-            self.lvalues.remove(item)
-        return self
-
-
-class DictVariable(Variable):
-    """A variable with a dictionnary of level:pretty name, where in the test "level" will be used but in display format "pretty name will be used"
-    for instance `ZC={1:Zero-copy,0:Copy}`
-
-    """
-    def __init__(self, name, data):
-        super().__init__(name)
-        if type(data) is dict:
-            self.vdict = data
-        else:
-            self.vdict = OrderedDict()
-            for g in data:
-                d = g.split(':')
-                self.vdict[d[0]] = d[1]
-
-    def makeValues(self):
-        return [(k, v) for k, v in self.vdict.items()]
-
-    def count(self):
-        return len(self.vdict)
-
-    def format(self):
-        k, v = next(self.vdict.items().__iter__())
-        return self.name, dtype(v)
-
-    def is_numeric(self):
-        k, v = next(self.vdict.items().__iter__())
-        return dtype(v) != str
-
-    def __add__(self, other):
-        self.vdict.update(other.vdict)
-        return self
-
-    def pop(self, item):
-        if item in self.vdict:
-            del self.vdict[item]
-        return self
-
-class RangeVariable(Variable):
-    def __init__(self, name, valuestart, valueend, log = False, step = None, force_int = False):
-        super().__init__(name)
-
-        if is_integer(valuestart) and is_integer(valueend):
-            valuestart=int(valuestart)
-            valueend=int(valueend)
-        else:
-            valuestart=float(valuestart)
-            valueend=float(valueend)
-        if valuestart > valueend:
-            self.a = valueend
-            self.b = valuestart
-        else:
-            self.a = valuestart
-            self.b = valueend
-        self.log = log
-        if step is None:
-            if log:
-                self.step = 2
-            else:
-                self.step = 1
-        else:
-            self.step = step
-        self.force_int = force_int
-
-    def count(self):
-        """todo: think"""
-        if self.step == "":
-            return len(FromFileVariable.getVals(self))
-        else:
-            if self.log:
-                return len(self.makeValues())
-            else:
-                return int(((self.b-self.a) / self.step) + 1)
-
-
-    def makeValues(self):
-            #Experimental design
-            if self.step == "":
-                vs =  self.a + (self.b-self.a) * FromFileVariable.getVals(self)
-            else:
-                vs = []
-                i = self.a
-                while i <= self.b:
-                    vs.append(i)
-                    if i == self.b:
-                        break
-                    if i == 0 and self.log:
-                        if self.b > 0:
-                            i = 1
-                        else:
-                            i = -1
-                    else:
-                        if self.log:
-                            i *= self.step
-                        else:
-                            i += self.step
-                if i > self.b:
-                    vs.append(self.b)
-            return [int(v) for v in vs] if self.force_int else vs
-
-    def format(self):
-        return self.name, int
-
-    def is_numeric(self):
-        return True
-
-class IfVariable(Variable):
-    """Condition variable.
-    
-    This is deprecated, prefer to use conditions in the test itself using a jinja template.
-    """
-    def __init__(self, name, cond, a, b):
-        super().__init__(name)
-        self.cond = cond
-        self.a = a
-        self.b = b
-
-    def makeValues(self):
-        vs = []
-        if aeval(self.cond):
-            return [self.a]
-        else:
-            return [self.b]
-
-    def count(self):
-        return 1
-
-    def format(self):
-        return str
-
-    def is_numeric(self):
-        return False
-
-class RandomVariable(Variable):
-    """
-    Generate a random number between 2 integers.
-    
-    This should generally be avoided.
-    """
-    def __init__(self, name, a, b):
-        super().__init__(name)
-        self.a = int(a.strip())
-        self.b = int(b.strip())
-
-    def makeValues(self):
-        return [random.randint(self.a, self.b)]
-
-
-    def count(self):
-        return 1
-
-    def format(self):
-        return self.name, int
-
-    def is_numeric(self):
-        return True
